Extract abstract parts of Repo/Adapter to Hexagram
This starts the process of extracting the parts of our
Repository/Adapter code that is not direcltly related to the Courses
service. Eventual plan is to put this in a gem. Right now I've gone as
far as moving the extracted code over to the Hexagram folder in /lib.

Tests that were only concerned with the extracted code have been moved
to tests in Hexagram.

The map_record_out code that we want to extract is not yet extracted. It
is still in CampusRepository and is being tested by that classes tests.
