--- app/models/adapters.rb
@@ -1,1 +1,0 @@
-module Adapters; end
