--- app/models/entities.rb
@@ -1,1 +1,0 @@
-module Entities; end
