--- app/models/models.rb
@@ -1,1 +1,0 @@
-module Models; end
