--- app/models/models/active_record.rb
@@ -1,3 +1,0 @@
-module Models
-  module ActiveRecord; end
-end
