+++ app/models/persisters/active_record/campus.rb
@@ -1,9 +1,9 @@
-module Models
+module Persisters
   module ActiveRecord
     class Campus < ::ActiveRecord::Base
 
       def self.orm_adapter
-        Adapters::ActiveRecord
+        ::Hexagram::Adapters::ActiveRecord
       end
 
       self.primary_key = "abbreviation"
