--- app/models/query_objects.rb
@@ -1,1 +1,0 @@
-module QueryObjects; end
