--- app/models/repositories.rb
@@ -1,1 +1,0 @@
-module Repositories; end
