+++ app/models/repositories/campus_repository.rb
@@ -1,10 +1,8 @@
-require_relative 'repository'
-
 module Repositories
   class CampusRepository
-    include Repository
+    include ::Hexagram::Repositories::Repository
 
-    def initialize(persistence_class = ::Models::ActiveRecord::Campus, orm_adapter = nil)
+    def initialize(persistence_class = ::Persisters::ActiveRecord::Campus, orm_adapter = nil)
       orm_adapter ||= persistence_class.orm_adapter
       super(persistence_class, orm_adapter)
     end
