+++ config/application.rb
@@ -12,6 +12,7 @@ Bundler.require(*Rails.groups)
 
 module CourseGuide
   class Application < Rails::Application
+    config.autoload_paths += %W(#{config.root}/lib/hexagram/lib/)
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
