+++ lib/hexagram/Rakefile
@@ -1,0 +1,5 @@
+require "rspec/core/rake_task"
+
+RSpec::Core::RakeTask.new(:spec)
+
+task default: [:spec]
