+++ lib/hexagram/lib/hexagram.rb
@@ -1,0 +1,5 @@
+module Hexagram
+end
+
+require_relative './hexagram/adapters.rb'
+require_relative './hexagram/repositories.rb'
