+++ lib/hexagram/lib/hexagram/adapters.rb
@@ -1,0 +1,4 @@
+module Hexagram
+  module Adapters; end
+end
+Dir.glob(File.join(File.dirname(__FILE__), "adapters", "*.rb")) { |file| require file }
