+++ lib/hexagram/lib/hexagram/repositories.rb
@@ -1,0 +1,4 @@
+module Hexagram
+  module Repositories; end
+end
+Dir.glob(File.join(File.dirname(__FILE__), "repositories", "*.rb")) { |file| require file }
