+++ lib/hexagram/spec/doubles/test_repository.rb
@@ -1,0 +1,8 @@
+class TestRepository
+  include ::Hexagram::Repositories::Repository
+end
+
+class TestEntity
+  def valid?(repo)
+  end
+end
