+++ lib/hexagram/spec/shared/adapter.rb
