--- spec/models/adapters/active_record_spec.rb
@@ -1,7 +1,0 @@
-require "rails_helper"
-require_relative "../../../app/models/adapters/active_record"
-require_relative "../../shared/adapter"
-
-RSpec.describe Adapters::ActiveRecord do
-  it_behaves_like "an adapter"
-end
