+++ spec/models/entities/campus_entity_spec.rb
@@ -1,5 +1,6 @@
 require "rails_helper"
 require_relative "../../../app/models/entities/campus_entity"
+require_relative "../../../app/models/repositories/campus_repository"
 
 RSpec.describe Entities::CampusEntity do
   let(:campus_instance) { described_class.new }
@@ -38,14 +39,14 @@ RSpec.describe Entities::CampusEntity do
 
   describe "valid?" do
     it "is valid if the repository says the abbrevation is unique" do
-      repository_double = instance_double(Repositories::CampusRepository)
+      repository_double = instance_double(CampusRepository)
       expect(repository_double).to receive(:unique?).with(abbreviation: "UMNTC").and_return(true)
 
       campus_instance.abbreviation = "UMNTC"
       expect(campus_instance.valid?(repository_double)).to be_truthy
     end
     it "is not valid if the repository says the abbrevation is not unique" do
-      repository_double = instance_double(Repositories::CampusRepository)
+      repository_double = instance_double(CampusRepository)
       expect(repository_double).to receive(:unique?).with(abbreviation: "UMNTC").and_return(false)
 
       campus_instance.abbreviation = "UMNTC"
